site/requirements.txt
pyyaml
requests
pelican[markdown]
invoke
livereload
watchdog
pelican-webassets
pelican-jinja2content
pelican-yaml-metadata
libsass==0.22.0 
packaging
